# backend/openapi.yaml
openapi: 3.0.3
info:
  title: SPV API
  version: 1.0.0
  description: API para votos, puntos y analisis de planificacion.
servers:
  - url: http://localhost:8787
paths:
...
  /api/achievements:
    get:
      summary: Listar logros votables por actividad
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: activityId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, voting, approved, rejected, archived]
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Logros y conteo de votos
    post:
      summary: Crear propuesta de logro votable
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [activityId, title]
              properties:
                activityId:
                  type: string
                title:
                  type: string
                description:
                  type: string
                rewardType:
                  type: string
                  enum: [points, item, mixed]
                rewardPoints:
                  type: integer
                rewardItemId:
                  type: string
                metadata:
                  type: object
      responses:
        '201':
          description: Logro creado en estado voting
  /api/achievements/{achievementId}/vote:
    post:
      summary: Votar un logro (up/down)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: achievementId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vote]
              properties:
                vote:
                  type: string
                  enum: [up, down]
                metadata:
                  type: object
      responses:
        '201':
          description: Voto registrado o actualizado
        '404':
          description: Logro no encontrado
        '409':
          description: Logro no esta en estado voting
  /api/achievements/{achievementId}/close:
    post:
      summary: Cerrar votacion de logro (approved/rejected/archived)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: achievementId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [approved, rejected, archived]
                note:
                  type: string
      responses:
        '200':
          description: Logro cerrado
        '404':
          description: Logro no encontrado
        '409':
          description: Logro ya cerrado o usuario no autorizado
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT